requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.35
alembic==1.13.2
python-dotenv==1.0.1
lxml==5.3.0
xmlschema==3.3.2
PyNaCl==1.5.0
requests==2.32.3
starlette==0.38.5
types-requests==2.32.0.20240914
prometheus-fastapi-instrumentator==7.0.0
web3==7.3.0
cachetools==5.3.3
openai==1.40.3
psycopg[binary]==3.2.3
rq==1.16.2
redis==5.0.8
pytest==8.3.4
